+++ app/src/test/java/com/techcourse/support/web/adaptor/HandlerAdaptorsTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.techcourse.support.web.mapping.ManualHandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -15,26 +14,6 @@ import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
 class HandlerAdaptorsTest {
 
     private final HandlerAdaptors handlerAdaptors = new HandlerAdaptors();
-
-    @Test
-    @DisplayName("handler 타입에 맞는 Adaptor를 반환한다.")
-    void getManualAdaptor() {
-        //given
-        HttpServletRequest request = mock(HttpServletRequest.class);
-        when(request.getRequestURI()).thenReturn("/login/view");
-        when(request.getMethod()).thenReturn("GET");
-
-        HandlerMapping manualHandlerMapping = new ManualHandlerMapping();
-        manualHandlerMapping.initialize();
-
-        Object handler = manualHandlerMapping.getHandler(request);
-
-        //when
-        HandlerAdaptor adaptor = handlerAdaptors.getAdaptor(handler);
-
-        //then
-        assertThat(adaptor).isInstanceOf(ManualHandlerAdaptor.class);
-    }
 
     @Test
     @DisplayName("handler 타입에 맞는 Adaptor를 반환한다.")
